add addMessage DB query

+++ db/queries/database.js
@@ -141,6 +141,24 @@ const checkUserIsAdmin = (userId) => {
     })
 };
 
+const addMessage = (messageInfo) => {
+  const queryString =`INSERT INTO messages (sender_id, recipient_id, content, item_id, date) VALUES ($1, $2, $3, $4, $5);`;
 
+  return db
+    .query(queryString, [messageInfo.userId, messageInfo.recipientId, messageInfo.content, messageInfo.item_id, messageInfo.date])
+    .then(() => {
+      return "Message saved succcessfuly!";
+    })
+}
 
-module.exports = { getItemsByCategory, getItemById, deleteItem, addItem, editItem, filterItems, getFavouriteItems, addFavoriteItem, getSellerInfo, checkUserIsAdmin };
+const getAdminByItemId = (item_id) => {
+  const queryString =  `SELECT owner_id FROM items WHERE id = $1;`;
+
+  return db
+    .query(queryString, [item_id])
+    .then((data) => {
+      return data.rows[0];
+    })
+}
+
+module.exports = { getItemsByCategory, getItemById, deleteItem, addItem, editItem, filterItems, getFavouriteItems, addFavoriteItem, getSellerInfo, checkUserIsAdmin, addMessage, getAdminByItemId };
